comment fix= arch: Alert/alertConfig.js, enums/constantes.js y helpers/colors.js
