+++ src/components/Alert/alertConfig.js
@@ -1,0 +1,6 @@
+export const iconsAlert = {
+  error: 'error',
+  info: 'info',
+  success: 'success',
+  warning: 'warning',
+};
