--- src/enums/alerts.js
@@ -1,29 +1,0 @@
-export const textAlert = {
-  aceptar: 'Aceptar',
-  confirmar: 'Confirmar',
-  aceptado: 'Aceptado',
-  cancelado: 'Cancelado',
-  cancelar: 'Cancelar',
-  confirmacion: 'Confirmación',
-  informacion: 'Información',
-  error: 'Error',
-  success: 'Éxito',
-  warning: 'Advertencia',
-  question: 'Pregunta',
-};
-
-export const colorAlert = {
-  confirm: '#00c853',
-  info: '#2196f3',
-  error: '#d50000',
-  success: '#00c853',
-  warning: '#ffc107',
-  question: '#ff9800',
-};
-
-export const iconsAlert = {
-  error: 'error',
-  info: 'info',
-  success: 'success',
-  warning: 'warning',
-};
