+++ src/helpers/colors.js
@@ -1,0 +1,8 @@
+export const colorAlert = {
+  confirm: '#00c853',
+  info: '#2196f3',
+  error: '#d50000',
+  success: '#00c853',
+  warning: '#ffc107',
+  question: '#ff9800',
+};
